// src/services/models/pageantdetails/pageant.ts
export interface Pageant {
  half_price_end_date_time_est?: any;
  half_price_start_date_time_est?: any;
  hide_votes_scheduler_date_time_est?: any;
  pca_start_date_time_est?: any;
  pca_end_date_time_est?: any;
  upcomingOrOngoingEventData?: any;
  id?: number;
  title?: string;
  slug?: string;
  master_pageant_id?: number;
  year_id?: number;
  age_from?: number;
  age_to?: number;
  start_date?: any;
  end_date?: any;
  allow_show_hide_event?: any;
  event_pb_hide_show_status?: any;
  gender?: string;
  is_married?: string;
  have_kids?: string;
  latitude?: any;
  longitude?: any;
  address?: any;
  website?: string;
  location?: any;
  venue?: any;
  email?: string;
  share_link?: string;
  phone?: any;
  description?: string;
  main_image?: string;
  main_image_full_url?: string;
  banner_image_full_url?: string;
  banner_image?: string;
  speciality_pageant?: string;
  show_recruit_director_button?: any;
  owner_id?: number;
  old_owner_id?: number;
  rating_average?: any;
  created_by_user_id?: number;
  updated_by_user_id?: number;
  type?: string;
  seo_focuskw?: any;
  seo_title?: any;
  seo_metakeyword?: any;
  seo_metadesc?: any;
  seo_linkdex?: any;
  seo_primary_age_divisions?: any;
  is_pageant_completed?: string;
  priority?: number;
  priority_weight?: number;
  priority_calculated?: number;
  facebook?: any;
  twitter?: any;
  status?: string;
  is_pca_activated?: string;
  people_award_name?: string;
  hide_vote_count?: number;
  per_vote_price?: number;
  view_count?: number;
  search_count?: number;
  cost_per_lead?: any;
  article_count?: any;
  active_membership_id?: any;
  active_membership_price?: any;
  is_verified?: string;
  pay_profile_id?: any;
  pca_start_date_time?: any;
  pca_end_date_time?: any;
  pca_admin_notify?: string;
  is_prediction_hide?: string;
  dnd_status?: string;
  dnd_inactive_date?: any;
  hide_message_me?: string;
  mail_notify_pref?: string;
  speciality_type?: number;
  txt_for_speciality_type?: any;
  streetAddress?: any;
  addressLocality?: any;
  addressRegion?: any;
  addressCountry?: any;
  redirect_membership_date?: any;
  half_price_start_notification_last_date?: any;
  import_id?: any;
  not_sure?: number;
  not_sure_end_date?: number;
  not_sure_address?: number;
  hide_votes_scheduler_date_time?: any;
  hide_contestant_last_name?: string;
  different_pca_end_dates?: string;
  contestants_sort_by_vote_count?: number;
  show_predictive_matrix?: number;
  one_winner_only?: number;
  is_half_price?: number;
  half_price_start_date_time?: any;
  half_price_end_date_time?: any;
  is_featured_event?: string;
  tickets_entry_fee_product_reminder_last_date?: any;
  create_event_reminder_last_date?: any;
  activate_pca_reminder_last_date?: any;
  event_timezone_id?: number;
  is_pca_deactivated_before_min_days?: number;
  days_left?: number;
  year_name?: YearName;
  pageant_awards?: PageantAward[];
  master_pageant?: Pageant;
  owner?: Owner;
  pageant_age_division?: AgeDivision[];
  pageant_contestant?: any[];
  pageant_country?: CountryState[];
  pageant_state?: CountryState[];
  pageant_phase_of_competition?: PageantPhaseOfCompetition[];
  pageant_sponsor?: any[];
  pageant_ad?: any[];
  events?: Pageant[];
  rating_count?: number;
  average_rating?: number;
  participants_count?: number;
  pageant_contestants?: Contestant[];
  pageantAgeDivisionsData?: AgeDivision[];
  pageantCountriesData?: CountryState[];
  pageantStatesData?: any[];
  pageantPhaseOfCompetitionsData?: PageantPhaseOfCompetition[];
  have_upcoming_events?: boolean;
  bip_awards_count?: number;
  event_result_count?: number;
  event_tense_status?: string;
  base_contestant_price_for_pca?: number;
  sub_speciality_type?: number;
  favourite?: any[];
  eventYearName?: number;
  pageant_id?: number;
  first_name?: string;
  last_name?: string;
  position?: string;
  quick_bio?: string;
  headshot_image?: string;
  headshot_image_full_url?: string;
  image_full_url?: string;
  business_title?: string;
  default_original_price?: number;
  default_discounted_price?: number;
  review_count?: number;
  role?: string;
  from_type?: string;
  from_tag_id?: number;
  to_type?: string;
  to_tag_id?: number;
  is_business?: any;
  business_role_id?: number;
  tagline?: any;
  image?: string;
  bio?: any;
  achievement?: any;
  facebook_link?: any;
  twitter_link?: any;
  instagram_link?: any;
  pinterest_link?: any;
  dob?: any;
  occupation?: any;
  achievements?: any;
  detail?: any;
  you_tube_link?: any;
  google_plus_link?: any;
  custom_design?: any;
  is_featured?: string;
  provide_coaching?: string;
  hourly_rate?: any;
  view_phone_count?: number;
  is_brand?: string;
  is_phase_available?: boolean;
  created_by?: any;
  updated_by?: any;
  created_at?: string;
  updated_at?: string;
}

export interface PageantPhaseOfCompetition {
  id: number;
  name: string;
  slug: string;
  master_record_type_id: number;
  description: string;
  image: any;
  status: string;
  have_images: boolean;
  pageant_id: number;
  competition_phase_id: number;
  phase_of_competition: PageantPhaseOfCompetition;
  title: string;
  main_image: string;
  banner_image: any;
  phone: number;
  email: string;
  website: any;
  speciality_type: number;
  sub_speciality_type: number;
  contestant_schedule_status: number;
}

export interface PageantAward {
  id: number;
  pageant_id: number;
  age_division_id: number;
  award_id: number;
  contestant_id: number;
  award: Award;
}